fix: unregister of service

+++ repositories/service-repository/src/main/kotlin/dev/redicloud/repository/service/ServiceRepository.kt
@@ -36,8 +36,7 @@ abstract class ServiceRepository<T : CloudService>(
                 val service = getService(serviceId) as T? ?: return@runBlocking
                 service.connected = false
                 if (service.currentSession() != null) service.endSession()
-                if (!service.unregisterAfterDisconnect()) return@runBlocking
-                if (service.canSelfUnregister()) {
+                if (service.canSelfUnregister() && service.unregisterAfterDisconnect()) {
                     registeredServices.remove(serviceId)
                     deleteService(transformShutdownable(service))
                 }else {
